CCS-457: Modularize co2_plume_tracking
