+++ src/ccs_scripts/co2_plume_extent/co2_plume_extent.py
@@ -4,30 +4,18 @@ Calculates the plume extent from a given coordinate, or well point,
 using SGAS and the dissolved property (AMFG/XMF2).
 """
 
-import logging
-import os
 import sys
 
-from ccs_scripts.co2_plume_extent.input import (
-    init_timer,
-    log_distance_calculation_configurations,
-    log_input_configuration,
-    make_parser,
-    replace_default_dummies_from_ert,
-    setup_log_configuration,
-)
 from ccs_scripts.co2_plume_extent.compute import calculate_distances
 from ccs_scripts.co2_plume_extent.config import (
-    Configuration,
     find_input_line,
     find_input_point,
 )
-from ccs_scripts.co2_plume_extent.output import (
-    collect_results_into_dataframe,
-    find_output_file,
-    log_results,
-    log_results_detailed,
+from ccs_scripts.co2_plume_extent.input import (
+    init_timer,
+    process_input,
 )
+from ccs_scripts.co2_plume_extent.output import export_results
 from ccs_scripts.utils.timer import Timer
 
 
@@ -51,22 +39,7 @@ def main():
     timer = Timer()
     timer.start("total")
 
-    args = make_parser().parse_args()
-    replace_default_dummies_from_ert(args)
-    args.column_name = (
-        args.column_name.upper() if args.column_name is not None else None
-    )
-    setup_log_configuration(args)
-    log_input_configuration(args)
-
-    config = Configuration(
-        args.config_plume_extent,
-        args.calc_type,
-        args.inj_point,
-        args.column_name,
-        args.case,
-    )
-    log_distance_calculation_configurations(config)
+    args, config = process_input()
 
     all_results = calculate_distances(
         args.case,
@@ -77,25 +50,8 @@ def main():
         args.threshold_dissolved,
     )
 
-    df = collect_results_into_dataframe(
-        all_results,
-        config,
-        config.injection_wells,
-    )
-    log_results(df)
-    log_results_detailed(df)
+    export_results(all_results, config, args.output_csv, args.case)
 
-    timer.start("export_results")
-    output_file = find_output_file(args.output_csv, args.case)
-    logging.info("\nExport results to CSV file")
-    logging.info(f"    - File path: {output_file}")
-    if os.path.isfile(output_file):
-        logging.info("Output CSV file already exists => Will overwrite existing file")
-    df.to_csv(output_file, index=False, na_rep="0.0")
-    timer.stop("export_results")
-
-    timer.stop("total")
-    timer.report()
     timer.stop("total")
     timer.report()
